Implement most of HIGHLIGHT, modularise PAINT
- Split term-counting, term scroll markers (term-marking), and
  jumping/stepping (term-walking) from PAINT (and also ELEMENT) into
  modules, distinguished and grouped by the new 'highlighting models'
  concept
- Simplify the AbstractEngine interface by incorporating the new classes
  introduced with these modules, via composition
- Hence simplify the PAINT and ELEMENT main modules
- Incorporate the new classes into HIGHLIGHT, causing nearly all
  features to be activated :tada:
  - Scroll markers are not quite activated, as they require an array of
    all actively highlighted elements, which is currently not readily
    accessible from HIGHLIGHT
