+++ src/modules/highlight/engines/element.mts
@@ -1,16 +1,16 @@
-import { type AbstractEngine, getMutationUpdates, containerBlockSelector } from "/dist/modules/highlight/engine.mjs";
+import { type AbstractEngine, getContainerBlock, getMutationUpdates } from "/dist/modules/highlight/engine.mjs";
 import { highlightTags } from "/dist/modules/highlight/highlighting.mjs";
 import { type AbstractSpecialEngine, DummySpecialEngine } from "/dist/modules/highlight/special-engine.mjs";
 import { PaintSpecialEngine } from "/dist/modules/highlight/special-engines/paint.mjs";
+import { StandardTermCounter } from "/dist/modules/highlight/models/tree-edit/term-counters/standard.mjs";
+import { StandardTermWalker } from "/dist/modules/highlight/models/tree-edit/term-walkers/standard.mjs";
+import { StandardTermMarker } from "/dist/modules/highlight/models/tree-edit/term-markers/standard.mjs";
 import { HIGHLIGHT_TAG, HIGHLIGHT_TAG_UPPER } from "/dist/modules/highlight/models/tree-edit/tags.mjs";
 import * as TermCSS from "/dist/modules/highlight/term-css.mjs";
 import type { MatchTerm } from "/dist/modules/match-term.mjs";
 import { requestCallFn } from "/dist/modules/call-requester.mjs";
-import {
-	EleID, EleClass, AtRuleID,
-	getNodeFinal, isVisible, getElementYRelative, elementsPurgeClass,
-	type TermHues, getTermClass, getTermToken,
-} from "/dist/modules/common.mjs";
+import type { UpdateTermStatus } from "/dist/content.mjs";
+import { EleID, EleClass, AtRuleID, elementsPurgeClass, type TermHues, getTermClass } from "/dist/modules/common.mjs";
 
 /**
  * Determines whether or not the highlighting algorithm should be run on an element.
@@ -21,33 +21,6 @@ import {
 const canHighlightElement = (rejectSelector: string, element: Element): boolean =>
 	!element.closest(rejectSelector) && element.tagName !== HIGHLIGHT_TAG_UPPER
 ;
-
-/**
- * Gets the containing block of an element.
- * This is its closest ancestor which has no tag name counted as `flow` in a highlight tags object.
- * @param element An element to find the first container block of (inclusive).
- * @returns The closest container block above the element (inclusive).
- */
-const getContainerBlock = (element: HTMLElement): HTMLElement =>
-	// Always returns an element since "body" is not a flow tag.
-	element.closest(containerBlockSelector) as HTMLElement
-;
-
-/**
- * Reverts the focusability of elements made temporarily focusable and marked as such using a class name.
- * Sets their `tabIndex` to -1.
- * @param root If supplied, an element to revert focusability under in the DOM tree (inclusive).
- */
-const elementsReMakeUnfocusable = (root: HTMLElement | DocumentFragment = document.body) => {
-	if (!root.parentNode) {
-		return;
-	}
-	root.parentNode.querySelectorAll(`.${EleClass.FOCUS_REVERT}`)
-		.forEach((element: HTMLElement) => {
-			element.tabIndex = -1;
-			element.classList.remove(EleClass.FOCUS_REVERT);
-		});
-};
 
 const ELEMENT_JUST_HIGHLIGHTED = "markmysearch__just_highlighted";
 
@@ -119,14 +92,30 @@ class FlowNodeList {
 }
 
 class ElementEngine implements AbstractEngine {
+	termOccurrences = new StandardTermCounter();
+	termWalker = new StandardTermWalker();
+	termMarkers = new StandardTermMarker();
+
 	mutationObserver: MutationObserver | null = null;
 	mutationUpdates = getMutationUpdates(() => this.mutationObserver);
 
 	specialHighlighter: AbstractSpecialEngine = new DummySpecialEngine();
 
-	constructor (terms: Array<MatchTerm>, hues: TermHues, updateTermStatus: (term: MatchTerm) => void) {
-		this.requestRefreshIndicators = requestCallFn(() => this.insertScrollMarkers(terms, hues), 50, 500);
-		this.requestRefreshTermControls = requestCallFn(() => terms.forEach(term => updateTermStatus(term)), 50, 500);
+	constructor (
+		terms: Array<MatchTerm>,
+		hues: TermHues,
+		updateTermStatus: UpdateTermStatus,
+	) {
+		this.requestRefreshIndicators = requestCallFn(() => (
+			this.termMarkers.insert(terms, hues, Array.from(document.body.querySelectorAll(terms
+				.slice(0, hues.length) // The scroll markers are indistinct after the hue limit, and introduce unacceptable lag by ~10 terms
+				.map(term => `${HIGHLIGHT_TAG}.${getTermClass(term.token)}`)
+				.join(", ")
+			) as NodeListOf<HTMLElement>))
+		), 50, 500);
+		this.requestRefreshTermControls = requestCallFn(() => (
+			terms.forEach(term => updateTermStatus(term))
+		), 50, 500);
 		this.mutationObserver = this.getMutationUpdatesObserver(terms);
 		this.specialHighlighter = new PaintSpecialEngine();
 	}
@@ -169,67 +158,6 @@ ${HIGHLIGHT_TAG} {
 	countMatches () {
 		this.requestRefreshIndicators?.next();
 		this.requestRefreshTermControls?.next();
-	}
-
-	insertScrollMarkers (terms: Array<MatchTerm>, hues: TermHues) {
-		if (terms.length === 0) {
-			return; // No terms results in an empty selector, which is not allowed.
-		}
-		const regexMatchTermSelector = new RegExp(`\\b${EleClass.TERM}(?:-\\w+)+\\b`);
-		const gutter = document.getElementById(EleID.MARKER_GUTTER) as HTMLElement;
-		const containersInfo: Array<{
-			container: HTMLElement
-			termsAdded: Set<string>
-		}> = [];
-		let markersHtml = "";
-		document.body.querySelectorAll(terms
-			.slice(0, hues.length) // The scroll markers are indistinct after the hue limit, and introduce unacceptable lag by ~10 terms
-			.map(term => `${HIGHLIGHT_TAG}.${getTermClass(term.token)}`)
-			.join(", ")
-		).forEach((highlight: HTMLElement) => {
-			const container = getContainerBlock(highlight);
-			const containerIdx = containersInfo.findIndex(containerInfo => container.contains(containerInfo.container));
-			const className = (highlight.className.match(regexMatchTermSelector) as RegExpMatchArray)[0];
-			const yRelative = getElementYRelative(container);
-			let markerCss = `top: ${yRelative * 100}%;`;
-			if (containerIdx !== -1) {
-				if (containersInfo[containerIdx].container === container) {
-					if (containersInfo[containerIdx].termsAdded.has(getTermToken(className))) {
-						return;
-					} else {
-						const termsAddedCount = containersInfo[containerIdx].termsAdded.size;
-						markerCss += `padding-left: ${termsAddedCount * 5}px; z-index: ${termsAddedCount * -1}`;
-						containersInfo[containerIdx].termsAdded.add(getTermToken(className));
-					}
-				} else {
-					containersInfo.splice(containerIdx);
-					containersInfo.push({ container, termsAdded: new Set([ getTermToken(className) ]) });
-				}
-			} else {
-				containersInfo.push({ container, termsAdded: new Set([ getTermToken(className) ]) });
-			}
-			markersHtml += `<div class="${className}" top="${yRelative}" style="${markerCss}"></div>`;
-		});
-		gutter.replaceChildren(); // Removes children, since inner HTML replacement does not for some reason
-		gutter.innerHTML = markersHtml;
-	}
-
-	raiseScrollMarker (term: MatchTerm | undefined, container: HTMLElement) {
-		const scrollMarkerGutter = document.getElementById(EleID.MARKER_GUTTER) as HTMLElement;
-		elementsPurgeClass(EleClass.FOCUS, scrollMarkerGutter);
-		[6, 5, 4, 3, 2].some(precisionFactor => {
-			const precision = 10**precisionFactor;
-			const scrollMarker = scrollMarkerGutter.querySelector(
-				`${term ? `.${getTermClass(term.token)}` : ""}[top^="${
-					Math.trunc(getElementYRelative(container) * precision) / precision
-				}"]`
-			) as HTMLElement | null;
-			if (scrollMarker) {
-				scrollMarker.classList.add(EleClass.FOCUS);
-				return true;
-			}
-			return false;
-		});
 	}
 
 	startHighlighting (
@@ -277,7 +205,7 @@ ${HIGHLIGHT_TAG} {
 		}
 		elementsPurgeClass(EleClass.FOCUS_CONTAINER, root);
 		elementsPurgeClass(EleClass.FOCUS, root);
-		elementsReMakeUnfocusable(root);
+		this.termWalker.cleanup();
 	}
 
 	/**
@@ -473,228 +401,12 @@ ${HIGHLIGHT_TAG} {
 		};
 	})();
 
-	focusNextTerm (
-		reverse: boolean,
-		stepNotJump: boolean,
-		term: MatchTerm | undefined,
-	) {
-		if (stepNotJump) {
-			// Currently no support for specific terms.
-			this.focusNextTermStep(reverse);
-		} else {
-			this.focusNextTermJump(reverse, term);
+	stepToNextOccurrence (reverse: boolean, stepNotJump: boolean, term?: MatchTerm | undefined): HTMLElement | null {
+		const focus = this.termWalker.step(reverse, stepNotJump, term);
+		if (focus) {
+			this.termMarkers.raise(term, getContainerBlock(focus));
 		}
-	}
-
-	/**
-	 * Focuses an element, preventing immediate scroll-into-view and forcing visible focus where supported.
-	 * @param element An element.
-	 */
-	focusElement (element: HTMLElement) {
-		element.focus({
-			preventScroll: true,
-			focusVisible: true, // Very sparse browser compatibility
-		} as FocusOptions);
-	}
-
-	// TODO document
-	selectNextElement (
-		reverse: boolean,
-		walker: TreeWalker,
-		walkSelectionFocusContainer: { accept: boolean },
-		elementToSelect?: HTMLElement,
-	): { elementSelected: HTMLElement | null, container: HTMLElement | null } {
-		const nextNodeMethod = reverse ? "previousNode" : "nextNode";
-		let elementTerm = walker[nextNodeMethod]() as HTMLElement;
-		if (!elementTerm) {
-			let nodeToRemove: Node | null = null;
-			if (!document.body.lastChild || document.body.lastChild.nodeType !== Node.TEXT_NODE) {
-				nodeToRemove = document.createTextNode("");
-				document.body.appendChild(nodeToRemove);
-			}
-			walker.currentNode = (reverse && document.body.lastChild)
-				? document.body.lastChild
-				: document.body;
-			elementTerm = walker[nextNodeMethod]() as HTMLElement;
-			if (nodeToRemove) {
-				nodeToRemove.parentElement?.removeChild(nodeToRemove);
-			}
-			if (!elementTerm) {
-				walkSelectionFocusContainer.accept = true;
-				elementTerm = walker[nextNodeMethod]() as HTMLElement;
-				if (!elementTerm) {
-					return { elementSelected: null, container: null };
-				}
-			}
-		}
-		const container = getContainerBlock(elementTerm.parentElement as HTMLElement);
-		container.classList.add(EleClass.FOCUS_CONTAINER);
-		elementTerm.classList.add(EleClass.FOCUS);
-		elementToSelect = Array.from(container.getElementsByTagName(HIGHLIGHT_TAG))
-			.every(thisElement => getContainerBlock(thisElement.parentElement as HTMLElement) === container)
-			? container
-			: elementTerm;
-		if (elementToSelect.tabIndex === -1) {
-			elementToSelect.classList.add(EleClass.FOCUS_REVERT);
-			elementToSelect.tabIndex = 0;
-		}
-		this.focusElement(elementToSelect);
-		if (document.activeElement !== elementToSelect) {
-			const element = document.createElement("div");
-			element.tabIndex = 0;
-			element.classList.add(EleClass.REMOVE);
-			elementToSelect.insertAdjacentElement(reverse ? "afterbegin" : "beforeend", element);
-			elementToSelect = element;
-			this.focusElement(elementToSelect);
-		}
-		if (document.activeElement === elementToSelect) {
-			return { elementSelected: elementToSelect, container };
-		}
-		return this.selectNextElement(reverse, walker, walkSelectionFocusContainer, elementToSelect);
-	}
-
-	/**
-	 * Scrolls to and focuses the next block containing an occurrence of a term in the document, from the current selection position.
-	 * @param reverse Indicates whether elements should be tried in reverse, selecting the previous term as opposed to the next.
-	 * @param term A term to jump to. If unspecified, the next closest occurrence of any term is jumpted to.
-	 */
-	focusNextTermJump (reverse: boolean, term?: MatchTerm) {
-		const termSelector = term ? getTermClass(term.token) : undefined;
-		const focusBase = document.body
-			.getElementsByClassName(EleClass.FOCUS)[0] as HTMLElement;
-		const focusContainer = document.body
-			.getElementsByClassName(EleClass.FOCUS_CONTAINER)[0] as HTMLElement;
-		const selection = document.getSelection();
-		const activeElement = document.activeElement;
-		if (activeElement && activeElement.tagName === "INPUT" && activeElement.closest(`#${EleID.BAR}`)) {
-			(activeElement as HTMLInputElement).blur();
-		}
-		const selectionFocus = selection && (!activeElement
-			|| activeElement === document.body || !document.body.contains(activeElement)
-			|| activeElement === focusBase || activeElement.contains(focusContainer)
-		)
-			? selection.focusNode
-			: activeElement ?? document.body;
-		if (focusBase) {
-			focusBase.classList.remove(EleClass.FOCUS);
-			elementsPurgeClass(EleClass.FOCUS_CONTAINER);
-			elementsReMakeUnfocusable();
-		}
-		const selectionFocusContainer = selectionFocus
-			? getContainerBlock(
-				selectionFocus.nodeType === Node.ELEMENT_NODE || !selectionFocus.parentElement
-					? selectionFocus as HTMLElement
-					: selectionFocus.parentElement,
-			) : undefined;
-		const walkSelectionFocusContainer = { accept: false };
-		const walker = document.createTreeWalker(document.body, NodeFilter.SHOW_ELEMENT, (element: HTMLElement) =>
-			element.tagName === HIGHLIGHT_TAG_UPPER
-			&& (termSelector ? element.classList.contains(termSelector) : true)
-			&& isVisible(element)
-			&& (getContainerBlock(element) !== selectionFocusContainer || walkSelectionFocusContainer.accept)
-				? NodeFilter.FILTER_ACCEPT
-				: NodeFilter.FILTER_SKIP);
-		walker.currentNode = selectionFocus ? selectionFocus : document.body;
-		const { elementSelected, container } = this.selectNextElement(reverse, walker, walkSelectionFocusContainer);
-		if (!elementSelected || !container) {
-			return;
-		}
-		elementSelected.scrollIntoView({ behavior: "smooth", block: "center" });
-		if (selection) {
-			selection.setBaseAndExtent(elementSelected, 0, elementSelected, 0);
-		}
-		document.body.querySelectorAll(`.${EleClass.REMOVE}`).forEach((element: HTMLElement) => {
-			element.remove();
-		});
-		this.raiseScrollMarker(term, container);
-	}
-
-	getSiblingHighlightFinal (
-		highlight: HTMLElement,
-		node: Node,
-		nextSiblingMethod: "nextSibling" | "previousSibling"
-	) {
-		return node[nextSiblingMethod]
-			? (node[nextSiblingMethod] as Node).nodeType === Node.ELEMENT_NODE
-				? (node[nextSiblingMethod] as HTMLElement).tagName === HIGHLIGHT_TAG_UPPER
-					? this.getSiblingHighlightFinal(node[nextSiblingMethod] as HTMLElement, node[nextSiblingMethod] as HTMLElement,
-						nextSiblingMethod)
-					: highlight
-				: (node[nextSiblingMethod] as Node).nodeType === Node.TEXT_NODE
-					? (node[nextSiblingMethod] as Text).textContent === ""
-						? this.getSiblingHighlightFinal(highlight, node[nextSiblingMethod] as Text, nextSiblingMethod)
-						: highlight
-					: highlight
-			: highlight;
-	}
-
-	getTopLevelHighlight (element: Element) {
-		const closestHighlight = (element.parentElement as Element).closest(HIGHLIGHT_TAG);
-		return closestHighlight ? this.getTopLevelHighlight(closestHighlight) : element;
-	}
-
-	stepToElement (element: HTMLElement) {
-		element = this.getTopLevelHighlight(element);
-		const elementFirst = this.getSiblingHighlightFinal(element, element, "previousSibling");
-		const elementLast = this.getSiblingHighlightFinal(element, element, "nextSibling");
-		(getSelection() as Selection).setBaseAndExtent(elementFirst, 0, elementLast, elementLast.childNodes.length);
-		element.scrollIntoView({ block: "center" });
-		this.raiseScrollMarker(undefined, getContainerBlock(element));
-	}
-
-	/**
-	 * Scrolls to and focuses the next occurrence of a term in the document, from the current selection position.
-	 * @param reverse Indicates whether elements should be tried in reverse, selecting the previous term as opposed to the next.
-	 * @param nodeStart __Only supplied in recursion.__ Specifies a node at which to begin scanning.
-	 */
-	focusNextTermStep (reverse: boolean, nodeStart?: Node) {
-		elementsPurgeClass(EleClass.FOCUS_CONTAINER);
-		elementsPurgeClass(EleClass.FOCUS);
-		const selection = getSelection();
-		const bar = document.getElementById(EleID.BAR);
-		if (!selection || !bar) {
-			return;
-		}
-		if (document.activeElement && bar.contains(document.activeElement)) {
-			(document.activeElement as HTMLElement).blur();
-		}
-		const nodeBegin = reverse ? getNodeFinal(document.body) : document.body;
-		const nodeSelected = reverse ? selection.anchorNode : selection.focusNode;
-		const nodeFocused = document.activeElement
-			? (document.activeElement === document.body || bar.contains(document.activeElement))
-				? null
-				: document.activeElement as HTMLElement
-			: null;
-		const nodeCurrent = nodeStart ?? (nodeSelected
-			? nodeSelected
-			: nodeFocused ?? nodeBegin);
-		if (document.activeElement) {
-			(document.activeElement as HTMLElement).blur();
-		}
-		const walker = document.createTreeWalker(document.body, NodeFilter.SHOW_ELEMENT, (element: HTMLElement) =>
-			(element.parentElement as Element).closest(HIGHLIGHT_TAG)
-				? NodeFilter.FILTER_REJECT
-				: (element.tagName === HIGHLIGHT_TAG_UPPER && isVisible(element))
-					? NodeFilter.FILTER_ACCEPT
-					: NodeFilter.FILTER_SKIP
-		);
-		walker.currentNode = nodeCurrent;
-		const element = walker[reverse ? "previousNode" : "nextNode"]() as HTMLElement | null;
-		if (!element) {
-			if (!nodeStart) {
-				this.focusNextTermStep(reverse, nodeBegin);
-			}
-			return;
-		}
-		this.stepToElement(element);
-	}
-
-	// Increasingly inaccurate as highlights elements are more often split.
-	getTermOccurrenceCount (term: MatchTerm) {
-		const occurrences = document.body.getElementsByClassName(getTermClass(term.token));
-		//const matches = occurrences.map(occurrence => occurrence.textContent).join("").match(term.pattern);
-		//return matches ? matches.length : 0; // Works poorly in situations such as matching whole words.
-		return occurrences.length; // Poor and changeable heuristic, but so far the most reliable efficient method.
+		return focus;
 	}
 
 	getMutationUpdatesObserver (terms: Array<MatchTerm>) {
